refactor(step3): Deck이 DrawStrategy를 주입받게 변경
