+++ src/main/kotlin/blackjack/BlackJackApplication.kt
@@ -3,6 +3,7 @@ package blackjack
 import blackjack.domain.BlackJackGame
 import blackjack.domain.Deck
 import blackjack.domain.PlayerResult
+import blackjack.domain.RandomDrawStrategy
 import blackjack.view.askHitFromPlayer
 import blackjack.view.getPlayerNames
 import blackjack.view.notifyDealerHit
@@ -11,7 +12,7 @@ import blackjack.view.printPlayerResults
 import blackjack.view.printResult
 
 fun main() {
-    val game = BlackJackGame(Deck())
+    val game = BlackJackGame(Deck(RandomDrawStrategy()))
     var dealer = game.getDealer()
     val players = game.dealWith(getPlayerNames())
     notifyStartGame(dealer, players)
