+++ src/main/kotlin/blackjack/domain/Deck.kt
@@ -1,27 +1,13 @@
 package blackjack.domain
 
-import java.util.LinkedList
-import java.util.Queue
-
-class Deck : DrawStrategy {
-    private val cards: Queue<Card> = LinkedList<Card>()
-
-    init {
-        cards.addAll(Card.ALL.shuffled())
+class Deck(
+    private val drawStrategy: DrawStrategy
+) {
+    fun fetchCard(): Card {
+        return drawStrategy.fetchCard()
     }
 
-    override fun fetchCard(): Card {
-        checkCardsAreEmpty()
-        return cards.poll()
-    }
-
-    private fun checkCardsAreEmpty() {
-        if (this.cards.isEmpty()) {
-            cards.addAll(Card.ALL.shuffled())
-        }
-    }
-
-    override fun getDealCards(): List<Card> {
-        return (1..DrawStrategy.DEAL_DRAW_COUNT).map { fetchCard() }
+    fun getDealCards(): List<Card> {
+        return (1..DrawStrategy.DEAL_DRAW_COUNT).map { drawStrategy.fetchCard() }
     }
 }
