+++ src/main/kotlin/blackjack/domain/DrawStrategy.kt
@@ -4,8 +4,6 @@ interface DrawStrategy {
 
     fun fetchCard(): Card
 
-    fun getDealCards(): List<Card>
-
     companion object {
         const val DEAL_DRAW_COUNT = 2
     }
